get rid entirely of wrapping for immutable objects (nil, true, false, ints, longs, chars)

+++ trufflesqueak/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/context/FrameSlotReadNode.java
@@ -29,7 +29,7 @@ public abstract class FrameSlotReadNode extends FrameSlotNode {
     }
 
     @Specialization(guards = "isInt(frame)")
-    public long readInt(VirtualFrame frame) {
+    public int readInt(VirtualFrame frame) {
         return FrameUtil.getIntSafe(frame, slot);
     }
